Update 29.03.17
 - Began setting up TrainAndTest exercise activity
 - Ensured that all callbacks in ACTRModel execute on the UI thread
 - Upgraded database to version 6, adding an index on the ParseId column
 - Allowed for querying of words by Parse ID, utilizing this inside JSONHandler.parseWord()
 - Added methods to ACTRModel and RequestManager for safe deconstruction
 - Added buildMultiError() to NetworkError for better reporting of multiple network errors with a single error object
 - Implemented functionality for tracking and storing sessions, and deleting unfinished sessions.
 - Implemented cleanSessions() in ACTRModel to ensure sessions conform to appropriate criteria
 - Made session link mandatory when constructing an interaction with InterxBuilder
 - Fixed issue with database not fully updating after pulling all Parse IDs
 - Modified timeout and retry parameters for network communications in RequestManager
 - Created LoadingBarView, and implemented inside TrainAndTest
 - Adjusted LearnModel.getNextWord() and related methods to accept Collections in addition to arrays
 - Implemented displaying of words and sentences in TrainAndTest
 - Improved list and map parsing utilities by utilizing JSON interpreters
 - Added ExerciseUtils to house utility methods shared by exercises
 - Implemented multiple presentation styles in TrainAndTest (training and testing)
 - Began implementation of tracking and storing interactions

+++ app/src/main/java/com/example/kathrin1/vokabeltrainer_newlayout/objects/InterxObject.java
@@ -29,7 +29,7 @@ public class InterxObject
     private final String result;
     private final int charCount;
     private final String exerciseType;
-    private Integer session;
+    private Long session;
 
     private Float activation; // Used for storing activation during activation calculations,
     // but is not stored in database.
@@ -172,9 +172,9 @@ public class InterxObject
         String exerciseType = cursor.getString(cursor.getColumnIndexOrThrow(DBHandler.INTERX_EXERCISE_TYPE));
 
         // Extract the session reference
-        Integer sessionId = cursor.isNull(cursor.getColumnIndexOrThrow(DBHandler.INTERX_SESSION))
-                            ? null
-                            : cursor.getInt(cursor.getColumnIndexOrThrow(DBHandler.INTERX_SESSION));
+        Long sessionId = cursor.isNull(cursor.getColumnIndexOrThrow(DBHandler.INTERX_SESSION))
+                         ? null
+                         : cursor.getLong(cursor.getColumnIndexOrThrow(DBHandler.INTERX_SESSION));
 
 
         // Build the interaction object
@@ -222,7 +222,7 @@ public class InterxObject
         return exerciseType;
     }
 
-    public Integer getSessionId()
+    public Long getSessionId()
     {
         return session;
     }
@@ -234,7 +234,7 @@ public class InterxObject
      * @param session The session to reference.  May be left null.
      * @return This InterxObject.
      */
-    public InterxObject setSessionId(Integer session)
+    public InterxObject setSessionId(Long session)
     {
         this.session = session;
         return this;
